[0.5.2] fix Canvas::clearRect

+++ example/Example.hpp
@@ -9,7 +9,7 @@ public:
 
     void onDraw();
     void onMouse(const glm::vec2& mouse);
-    void onResize(const glm::vec2& size);
+    void onResize(const glm::ivec2& size);
     void onKey(int k, int a, int m);
 private:
     canvas2D::Canvas canvas;
